protoc 파일 작동
작동된 후 다시 proto파일 read후 template파일에 있는거 읽어서 해당하는 핸들러 파일 만들기
핸들러 파일 만든 후 필요한 파일들 있어야 하는 디렉토리로 이동
필요없는 부수 파일들 삭제 처리
까지 자동화

+++ ServerCore/Session.cpp
@@ -246,7 +246,7 @@ void Session::ProcessRecv(const int32 numOfBytes)
 	const uint32 dataSize = m_RecvBuffer.DataSize();
 
 	if (const int32 processLen = OnRecv(m_RecvBuffer.ReadPos(), dataSize); 
-		processLen < 0 || dataSize < processLen || m_RecvBuffer.OnRead(processLen) == false)
+		processLen < 0 || static_cast<int32>(dataSize) < processLen || m_RecvBuffer.OnRead(processLen) == false)
 	{
 		Disconnect(L"OnRead Error");
 		return;
